Upgrade all dependencies to newest versions
There was some dependency issues between the version of
angulartics that this project depended on, and the version
angulartics-google-analytics dependend on which made installing the
dependencies require manual resolution (which made installation through
grunt fail).

It would have sufficed to upgrade angulartics, but while I were at it, I
took all of the other dependencies as well. Everything seems to work
fine with the new versions.