Support showing videos
This traverses all the files that should be shown, and adds a property
"type" to them. The type is either "video" if the file ends with ".webm"
or "img" otherwise.

If the type is "img", an img tag is used as before. However, if the type
is "video", a video tag is used instead. The attributes autoplay and
loop are included so they appear similarly to animated gif files.

Currently, video tags with multiple sources are not supported, so the
video will have to be either .webm or .mp4 (if both are added, they will
appear as two separate videos). This means that depending on the format
chosen, the video may not be playable on all devices.