Add videos that were removed in 3e557f5
Now that videos are supported, these can be re-added. They are added in
the .webm format, instead of the .gifv format that were previously used.
This is because the .gifv format is a html wrapper around the video to
provide multiple formats.

The .gifv format is not supported, because it would have to be included
as an iframe. With an iframe, the size of the iframe has to be set to
the size of the content, which is inconvenient, and not always possible.
If the size is not set, it would get a default size which in most cases
would be incorrect.